aarch64: Fix memory cost for FPM_REGNUM
GCC 15 is going to be the first release to support FPMR.
While working on a follow-up patch, I noticed that for:

    (set (reg:DI R) ...)
    ...
    (set (reg:DI fpmr) (reg:DI R))

IRA would prefer to spill R to memory rather than allocate a GPR.
This is because the register move cost for GENERAL_REGS to
MOVEABLE_SYSREGS is very high:

  /* Moves to/from sysregs are expensive, and must go via GPR.  */
  if (from == MOVEABLE_SYSREGS)
    return 80 + aarch64_register_move_cost (mode, GENERAL_REGS, to);
  if (to == MOVEABLE_SYSREGS)
    return 80 + aarch64_register_move_cost (mode, from, GENERAL_REGS);

but the memory cost for MOVEABLE_SYSREGS was the same as for
GENERAL_REGS, making memory much cheaper.

Loading and storing FPMR involves a GPR temporary, so the cost should
account for moving into and out of that temporary.

This did show up indirectly in some of the existing asm tests,
where the stack frame allocated 16 bytes for callee saves (D8)
and another 16 bytes for spilling a temporary register.

It's possible that other registers need the same treatment
and it's more than probable that this code needs a rework.
None of that seems suitable for stage 4 though.

gcc/
	* config/aarch64/aarch64.cc (aarch64_memory_move_cost): Account
	for the cost of moving in and out of GENERAL_SYSREGS.

gcc/testsuite/
	* gcc.target/aarch64/acle/fpmr-5.c: New test.
	* gcc.target/aarch64/sve2/acle/asm/dot_lane_mf8.c: Don't expect
	a spill slot to be allocated.
	* gcc.target/aarch64/sve2/acle/asm/mlalb_lane_mf8.c: Likewise.
	* gcc.target/aarch64/sve2/acle/asm/mlallbb_lane_mf8.c: Likewise.
	* gcc.target/aarch64/sve2/acle/asm/mlallbt_lane_mf8.c: Likewise.
	* gcc.target/aarch64/sve2/acle/asm/mlalltb_lane_mf8.c: Likewise.
	* gcc.target/aarch64/sve2/acle/asm/mlalltt_lane_mf8.c: Likewise.
	* gcc.target/aarch64/sve2/acle/asm/mlalt_lane_mf8.c: Likewise.