c++, contracts, Darwin: Fix local symbols for linker.
Darwin partitions objects into 'atoms' based on linker-visible
symbols.  If these are not present, and weak and non-weak sections
overlap we get spurious warnings.  We need the contract violation
constants and the source locations to be visible.

In addition, this fixes those symbols such that we do not
unconditionally emit periods into the names (some assemblers do
not accept this).

TODO: fix other cases at some stage.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>