Remove unnecessary `record` argument from maybe_version_functions.
Previously, the `record` argument in maybe_version_function allowed the
call to cgraph_node::record_function_versions to be skipped.  However,
this was only skipped when both decls were already marked as versioned,
in which case we trigger the early exit in record_function_versions
instead. Therefore, the argument is unnecessary.

gcc/cp/ChangeLog:

	* class.cc (add_method): Remove argument.
	* cp-tree.h (maybe_version_functions): Ditto.
	* decl.cc (decls_match): Ditto.
	(maybe_version_functions): Ditto.