testsuite: Use int size instead of alignment for pr116357.c
The test case assumes that alignof(int)=sizeof(int).  But for some
targets this is not valid.  For example, for PRU target,
alignof(int)=1 but sizeof(int)=4.

Fix the test case to align to twice the size of int, as the expected
dg-error messages suggest.

This patch fixes the test failures for PRU target.

gcc/testsuite/ChangeLog:

	* gcc.dg/pr116357.c: Use sizeof(int) instead of alignof(int).

Signed-off-by: Dimitar Dimitrov <dimitar@dinux.eu>