Darwin: Update handling of long constant names.
The long(er) names we are using for contracts constants (e.g.
Lcontract_violation) is triggering a warning where the prefix
to ASM_GENERATE_INTERNAL_LABEL is shorter than our literal;
that warning becomes an error after stage1 and so needs to be
fixed to allow bootstrap on Darwin.

This needs to be fixed more generally for Darwin upstream but
the patch here is sufficient for the contracts branch and stage

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>