Prune non-SLP paths from vectorizer loop analysis
The following prunes non-SLP iteration and the parts of non-SLP
stmt analysis that is no longer necessary - we need to keep the
parts that bail on stmts not covered by SLP discovery or that
failed SLP discovery.  This will only go away when a we can build
a fully covering single-lane SLP graph to fall back to.

	* tree-vect-loop.cc (vect_analyze_loop_operations): Prune
	all actual analysis and only fail when we discover a not
	SLP covered stmt.
	(vect_analyze_loop_2): Remove path trying without SLP.